update toolkit route
